requirements.txt
firebase-admin==6.9.0
streamlit==1.51.0
altair==5.5.0
pandas==2.2.2
numpy==2.0.1
google-auth==2.35.0
